p2b: using the new [encode/decode]AsNameElement feature to ensure we can support arbitrary plugin names, which fixes remote plugins not working issue.
Slashes are reserved characters in Vanadium names and they have special
meaning. In P2B the name/address of a plugins can be anything specially
remote plugins which are urls and have / in them.

Now we expect the name of the plguin to be encoded as a name part which
percent encodes / in the name and we decode on the receiving side.

Change-Id: Ic60613d3d29937535aa04fb141493a17692ea6ad